fix: Modals were not activating in settings page
This commit refactors the UI to rely on a common `<WalletLayout`
component.  There was functionality inside of the wallet index that was
not available in the settings ui.  This commit also refactors the modal
and layout components out of the `@/views/Wallet` directory and moves
them to `@/components/{layout,modals}` directories for clarity.
