+++ src/components/layout/WalletLoading.vue
