+++ src/components/layout/WalletNavLink.vue
