+++ src/components/layout/WalletSidebar.vue
