+++ src/components/layout/WalletSidebarDefault.vue
