+++ src/components/modals/WalletErrorModal.vue
