+++ src/components/modals/WalletHideAccountModal.vue
