+++ src/components/modals/WalletLedgerDeleteModal.vue
