+++ src/components/modals/WalletLedgerVerifyDecryptModal.vue
