+++ src/components/modals/WalletNewDeviceModal.vue
