+++ src/views/Home/index.vue
@@ -75,7 +75,7 @@ import { useRouter } from 'vue-router'
 import { useHomeModal, useWallet } from '@/composables'
 import { firstValueFrom } from 'rxjs'
 import { Network } from '@radixdlt/application'
-import WalletLoading from '@/views/Wallet/WalletLoading.vue'
+import WalletLoading from '@/components/layout/WalletLoading.vue'
 import { hasKeystore } from '@/actions/vue/create-wallet'
 import { getAcceptedTos, setAcceptedTos } from '@/actions/vue/data-store'
 import { defaultNetwork } from '@/helpers/network'
