+++ src/views/Settings/index.vue
@@ -51,7 +51,7 @@ import SettingsResetPassword from './SettingsResetPassword.vue'
 import SettingsSelectNode from './SettingsSelectNode.vue'
 import SettingsTokens from './SettingsTokens.vue'
 import SettingsSelectDecimal from './SettingsSelectDecimal.vue'
-import WalletLayout from '@/components/WalletLayout.vue'
+import WalletLayout from '@/components/layout/WalletLayout.vue'
 import { Ref, ref } from '@nopr3d/vue-next-rx'
 import { useSettingsTab, useWallet } from '@/composables'
 import { useRouter } from 'vue-router'
